Fix JFR event names in the flight recorder profile (#15506)
Motivation:
We updated the event names in https://github.com/netty/netty/pull/15500
to align with best practices, but forgot to update the JFR profile as
well.

Modification:
Add a set of new events for the new names.
Keep the old events around, in case people use this profile on Netty
4.2.3.

Result:
The profile template works again on the latest snapshot.