IoUring: Detect completion queue overflow (#15505)
Motivation:

If the completion queue is sized to low we might run into a state where
it would have overflown and so the kernel has to buffers completions
until there is space again. We should detect it, handle it and log so
users can increase the size.

Modifications:

Expose queue overflow state

Result:

Related to https://github.com/netty/netty/issues/15502