Add close-tracking to our leak detector (#15512)
Motivation:
Double-free is a common problem that people run into with reference
counting (and also without, but I digress), but our leak detector
currently only helps when systems make too few calls to
`ByteBuf.release` – not when they make too many.

Modification:
Add an option to track when resources are closed, in addition to their
life time. With this enabled, closing a resource will replace their
lifetime sequence with a special close-marking `TrackRecord` instead of
`null`. We can then attach this record as a suppressed exception to the
`IllegalReferenceCountException` thrown by `release`.

Result:
Exceptions from double-free bugs can now carry the stack trace of when
the object was actually released, making it easier to debug such issues.
This hooks into the "leak aware" buffer mechanism, so it's subject to
the same sampling and overhead reductions as the rest of our leak
tracking.