Do no check length parameter of ByteBufOutputStream.write(byte[], int, int); method
Motivation:

ByteBufOutputStream#write(byte[], int, int); method should respect the contract of the DataOutput interface and of the OutputStream class.
In particular, throwing an IndexOutOfBoundsException if offset is negative.

This is not done when the given length is 0. In this case, the method just return.

Modifications:

Removing the check of length parameter preventing enter in buffer.writeBytes(b, off, len);
Add a test, checking the method throws an IndexOutOfBoundsException when offset is negative even if the length parameter is 0.

Result:

The ByteBufOutputStream#write(byte[], int, int); method will throw an IndexOutOfBoundsException when offset is negative even if the length parameter is 0.