Merge commit from fork
* Enforce the maximum number of RST frames that can be sent in window of time

Motivation:

A remote peer might be able to trigger an instance to generate and send RST frames by sending invalid frames on an existing stream. This can cause high resource usage and so might be abused by a remote peer.

Modifications:

Limit the number of RSTs that we allow to be generated and so send in a specific time window. If this limit is reached a GO_AWAY frame is send and the connection be closed.

Result:

Fix high resource usage that can be caused by a remote peer by trigger RST frames

* Adjust testing

* Address comments