Create stream when unknown frame is received Motivation:
Receiving an unknown frame on a stream that is in 'idle' state (i.e. not created) would produce a NullPointerException.

Modification:

If an unknown frame is received, create the associated stream if necessary. This is in accordance with [RFC 9113](https://httpwg.org/specs/rfc9113.html#StreamStates):

> Receipt of frames for which the semantics are unknown cannot be treated as an error, as the conditions for sending and receiving those frames are also unknown

Technically it would be better to use the frame without a stream, but Http2UnknownFrame needs a Http2FrameStream to access the stream at all.

Result:

No NPE, user can handle unknown frame as normal.

Found by fuzzing.