HTTP: Allow to configure HttpServerUpgradeHandler to remove itself after the first request on a connection
Motivation:

While it is not strictly required that the first request must be the upgrade request it is usually what is the case in real world and it is even required by some protocols. We should allow to configure the HttpServerUpgradeHandler to remove itself and so reduce overhead.

Modifications:

- Add new constructor which allows to specify that we remove the handler after the first request even if its not an upgrade
- Adjust unit test

Result:

Fixes https://github.com/netty/netty/issues/15489