Drop unknown frame on missing stream (#15592) (#15595)
Motivation:

Receiving an unknown frame on a stream that is in 'idle' state (i.e. not
created) would produce a NullPointerException.

Modification:

If an unknown frame is received and there is no associated stream, drop
that frame. This is in accordance with [RFC
9113](https://httpwg.org/specs/rfc9113.html#StreamStates):

> Receipt of frames for which the semantics are unknown cannot be
treated as an error, as the conditions for sending and receiving those
frames are also unknown

Another option would be to create the stream.

Result:

No NPE.

Found by fuzzing.

Co-authored-by: Jonas Konrad <me@yawk.at>