Only allow auto-scaling of EventLoop's when IoHandler supports it. (#15603)
Motivation:

We recently added the feature of auto-scaling event-loops to be able to
have a more elastic setup. While this works out in most cases some
transport implementations might not allow it as it is not possible to
change threads due performance reasons. One of these examples is atm
io_uring. Here we require that the thread will never change during the
life-time of the ring as we setup the ring with
IORING_SETUP_SINGLE_ISSUER for performance reasons. Changing the thread
results in have io_uring_enter fail with EEXIST.

Modifications:

- Add new default method to IoHandlerFactory which can be used to either
signal support or no support for changing the Thread during the
life-time of IoHandler instances that are created via the factory.
- Override the method for Nio, Epoll and Kqueue to signal that changing
is supported and so our current auto-scaling implementation can be used
for these.
- Don't override the method for IoUring as we can't support it with the
current setup.

Result:

No more EEXIST errors when using io_uring with auto-scaling. This is
related to https://github.com/netty/netty/pull/15524