Merge commit from fork (#15612)
Motivation:

We should ensure our decompressing decoders will fire their buffers
through the pipeliner as fast as possible and so allow the user to take
ownership of these as fast as possible. This is needed to reduce the
risk of OOME as otherwise a small input might produce a large amount of
data that can't be processed until all the data was decompressed in a
loop. Beside this we also should ensure that other handlers that uses
these decompressors will not buffer all of the produced data before
processing it, which was true for HTTP and HTTP2.

Modifications:

- Adjust affected decoders (Brotli, Zstd and ZLib) to fire buffers
through the pipeline as soon as possible
- Adjust HTTP / HTTP2 decompressors to do the same
- Add testcase.

Result:

Less risk of OOME when doing decompressing

Co-authored-by: yawkat <jonas.konrad@oracle.com>