Make ResourceLeakDetector.track not final Motivation:
For fuzzing, paranoid leak detection is necessary, but running GC for each test is too slow. Instead, I check the `allLeaks` field after each run for any open trackers (with some pruning to avoid static buffers). This means that a lot of the tracked data, and in particular the whole reference queue, is unnecessary.

For better performance, it would be nice to instead only track the number of open leaks, and compare that number before and after a test runs. This can be achieved using a custom ResourceLeakTracker implementation, but for that ResourceLeakDetector.track must become non-final.

Modification:

Make the method non-final.

Also fix a bunch of warnings.

Result:

ResourceLeakDetector.track can be overridden for more efficient leak detection.