LeakPresenceDetector draft Motivation:
For some tests such as Bzip2IntegrationTest.testHugeDecompress, normal leak detection is too slow. Leak *presence* detection can be very fast though – it just won't give us any information about the leak if it occurs.

Modification:

Introduce a LeakPresenceDetector that uses a LongAdder to count open resources. A LeakPresenceExtension checks that counter after each test, and reports an error if a resource remains open.

The big problem with this approach is that some resources are held in static variables, which may be initialized inside a test, and will count towards the counter. Fundamentally, the only way to exclude these resources involves capturing a stack trace during tracker creation and checking if we're in some `<clinit>`. We could avoid actually materializing the StackTraceElements, but that would mean saving the trackers somewhere where we can inspect them, which is also not ideal.

Instead, I chose to add a wrapper method `staticInitializer` that static initializers would need to wrap around any resource creation. This is demonstrated in `LeakPresenceDetectorTest.ResourceInStaticVariable1`. This approach is pretty performant, but it would require changes in a few places inside netty, and potentially in user code. Since this is only necessary for the improved LeakPresenceDetector which is intended for tests and opt-in, I think this is an acceptable compromise.

Alternative suggestions welcome.

Result:

Low-cost leak detection for tests. Bzip2IntegrationTest runs basically just as fast as without leak detection.

Leaks can be unambiguously associated with a particular test, instead of only being reported on GC.