Allow the OpenSSL implementation to use the java Signature for keys that cannot be obtained directly
 #### Motivation

Sometimes the private key is not accessible to the application directly
such as when using hardware based keys, key signature forwarding, etc.
Many of these keys are exposed to the application via implementations
of the java.security.Provider system. This means we are stuck using the
JDK engine or re-implementing the signing/decrypting algorithms.

 #### Modifications

Teach netty how to look up the Signature that corresponds to a PrivateKey
implementation and then use the existing mechanisms for external signing
and decryption.