BouncyCastleAlpnSslUtils needs to use the correct SSLEngine class as otherwise it will fail to init static fields.
Motivation:

We recently did some changes related to how BouncyCastle is loaded and while doing so introduce a regression which would
cause NPE's later on once BouncyCastle was detected and used and we tried to apply ALPN later on. This regression was introduced by
https://github.com/netty/netty/pull/15533

The problem was that we used the wrong class to try to obtain the Method instances via reflection. This failed as the class itself is package-private
and so did not allow to access the methods. To fix this we need to ensure we use the public interface that is used by BouncyCastle and not the implemention
class itself. Beside this we also should be more defensive and only try to use ALPN via BouncyCastle if we can obtain the Methods in the first place.

Modifications:

- Fix refelective access
- Add more guards
- Add unit test

Result:

Fixes https://github.com/netty/netty/issues/15627

Result:

After your change, what will change.