Teach the BoringSSL implementation to use a java.security.Signature for keys that cannot be obtained directly (#15626)
Motivation:

Sometimes the private key is not accessible to the application directly
such as when using hardware based keys, key signature forwarding, etc.
Many of these keys are exposed to the application via implementations of
the java.security.Provider system. This means we are stuck using the JDK
engine or re-implementing the signing/decrypting algorithms.

Modification:

Teach netty how to look up the Signature that corresponds to a
PrivateKey implementation and then use the existing mechanisms for
external signing and decryption.

---------

Co-authored-by: Norman Maurer <norman_maurer@apple.com>