Drop unknown frame on missing stream in first packet Motivation:
If an unknown frame with an unknown stream is received in the first packet, the PrefaceFrameListener is called instead of the normal frame listener. This means that the fix in #15592 is bypassed.

Modification:

Modify PrefaceFrameListener to require the preface has come in (this is checked by the decoder already) and delegate to the current listener.

Inline onUnknownFrame0 because it's only used in the normal listener, now.

Result:

No NPE.

Found by fuzzing