drop a bunch of unnessacary reflection (#15636)
Motivation:
With the bump of the main branch to java 22 a bunch of reflection and
java version guards are no longer needed as the relevant methods can be
used directly.

Modification:
Remove the relevant reflection calls java version guards.

Result:
No more unnessacary reflection and java version guards that just make
the code harder to read.