update/cleanup codeql workflow (#15683)
Motivation:
As stated when executing the workflow, the major version (v2) of codeql
that is currently in use has long been deprecated. This PR updates the
codeql actions to v3 and also includes bumping the other action version
majors (major bump due to a new major nodejs requirement, not relevant
on github-hosted runners) as well as a bit of cleanup to the action.

Modification:
Updates the codeql actions to v3, updates actions/checkout and
action/setup-java to v5 (codeql action only), cleanup of codeql workflow
definition.

Result:
The latest codeql version is in use and the workflow definition is
cleaned up.