Native: Include error value when throwing NativeIoException in the me… (#15704)
…… (#15700)

…ssage

Motivation:

We should include the error value (which is a negative errno) in the
message of the NativeIoException. This allows easier debugging as it is
much easier to lookup the errno value and its cause then the human
readable message.

Modification:

Include error in the message of the exception

Result:

Easier to debug failed native ops