Fix Snappy compression bug
Motivation:
Certain inputs to the Snappy compression routine can cause an IndexOutOfBoundsException to be thrown.

Modification:
Fix conversion of signed-short to unsigned 16-bit ints.
These numbers were used as indices, and negative values caused the exception to be thrown.
Add a data pattern to the inputs of the `testHugeDecompress` test (this input pattern catches the bug) and check that the decompressed outputs match the inputs, using a SHA-256 checksum.

Result:
The `testHugeDecompress` test now verifies round-trip data integrity, and we've fixed a bug in Snappy that this found.