Add new callbacks to IoHandle implementation to allow better handling of resources
Motivation:

b7fcf5785dd8fb72019d47565656cd39ae25246b added a new method to IoUringIoHandle which allowed it to better handle native resources. After thinking more about it we should better just move it to IoHandle directly and also add another method
that is called when an IoHandle was succesfully registered. This allows to allocate / deallocate resources in a consistent and safe way in any transport.

Modification:

- Move default method from IoUringIoHandle to IoHandle
- Add new default method to IoHandle which is called when a handle is registered successfully.
- Adjust all IoHandler implementations to call these methods correctly.

Result:

Better and more flexible implemention of IoHandle possible