bump netty5 java version requirement to 25 (#15676)
Motivation:
With the addition of MemorySegment into the JDK the Unsafe and
ByteBuffer based Buffer implementations are no longer needed. Due to a
bug on jdk versions before 25 the nio buffers created from memory
segments cannot be used for certain operations, for example compression,
making the Unsafe/ByteBuffer implementations a requirement when running
on versions 22-24 (with the downside that Unsafe will print warnings
upon use and the ByteBuffer lifetime not being controlled by netty).
Given these reasons it seems to make sense to bump to 25, which would
allow to remove the Unsafe/ByteBuffer implementations in a subsequent
step as the MemorySegment based buffer is all that's needed.

Modification:
Bump the jdk version used for building to 25 and the compile target to
25.

Result:
Netty 5 requires java 25 instead of 22.