Avoid buffer duplication in AdaptiveByteBuf.setBytes
Motivation:
The setBytes methods are mutating methods, and therefor not thread-safe.
The pooling allocator do not protect the position of its internal NIO buffer in these methods, which means it does not call duplicate on the buffer before using it.
The adaptive allocator can make the same optimization.

Modification:
Remove ByteBuffer.duplicate calls from the setBytes methods of the AdaptiveByteBuf.

Result:
No more unnecessary ByteBuffer allocation when calling these methods, so the performance should now be on par with the pooling allocator.

Fixes https://github.com/netty/netty/issues/15723