Allow CertificateBuilder to specify Provider instance to use (#15722)
Motivation:

The 4.2 release introduced a new module, `netty-pkitesting`, which
includes a `CertificateBuilder` class. It greatly simplifies the
certificate generation but one particular limitation that we run into is
that `CertificateBuilder` implementation only supports
`BouncyCastleProvider` (see please
https://github.com/netty/netty/blob/4.2/pkitesting/src/main/java/io/netty/pkitesting/Algorithms.java#L93).

In OpenSearch fe we are working towards supporting FIPS mode and use
`BouncyCastleFipsProvider` only (which Netty is also aware of, see
please
https://github.com/netty/netty/blob/4.2/handler/src/main/java/io/netty/handler/ssl/util/BouncyCastleUtil.java#L37).
The `CertificateBuilder` class would greatly simplify the testing
efforts but we could not use it with non-default provider.

Modification:

Allow `CertificateBuilder` to specify `Provider` instance to use.

Result:

```java
      CertificateBuilder builder = new CertificateBuilder()
                .algorithm(algorithm)
                .provider(new BouncyCastleJsseProvider());
```

---------

Signed-off-by: Andriy Redko <drreta@gmail.com>