Faster slicing of unpooled direct buffers (#15727)
Motivation:
The UnpooledDirectByteBuf is used as the chunk implementation in the
adaptive allocator. As a consequence, its `ByteBuf.nioBuffer(int, int)`
method is used to obtain the internal NIO buffers of the allocated
adaptive buffers.

Modification:
On Java 13, we can implemenet this in terms of the new
`ByteBuffer.slice(int, int)`, which allocates the slice directly and
avoids allocating a temporary duplicate buffer.

Result:
AdaptiveByteBuf method that rely on the internal NIO buffer are now
faster when they need to allocate the underlying ByteBuffer.

Fixes https://github.com/netty/netty/issues/15723