IoUring: Verify configuration early and ensure its immutable. (#15729)
Motivation:

We should do as much validation as possible early and so fail fast when
creating the IoHandleFactory itself. Beside this we also need to ensure
things are immutable once we pass it the factory and can't be changed by
the user.

Modification:

- Do the validation early
- Act on an immutable config

Result:

Fail fast and guard from users messing up config once the
IoHandlerFactory is created