Faster leak detector implementation for tests (#15622)
Motivation:

For some tests such as Bzip2IntegrationTest.testHugeDecompress, normal
leak detection is too slow. Similarly, for fuzzing, I have to use
paranoid leak detection to detect bugs in a reproducible way, which in
turn adds a performance cost that may prevent some bugs from being
found.

Leak *presence* detection could be very fast though – it just won't give
us any details about the leak when it triggers.

Modification:

Introduce a LeakPresenceDetector (netty-common) that uses a LongAdder to
count open resources. A LeakPresenceExtension (netty-testsuite-common,
not public API) checks that counter after each test, and reports an
error if a resource remains open.

The big problem with this approach is that some resources are held in
static variables, which may be initialized inside a test, and will count
towards the counter. Fundamentally, the only way to exclude these
resources involves capturing a stack trace during tracker creation and
checking if we're in some `<clinit>`. We could avoid actually
materializing the StackTraceElements, but that would mean saving the
trackers somewhere where we can inspect them, which is also not ideal.

Instead, I chose to add a wrapper method `staticInitializer` that static
initializers would need to wrap around any resource creation. This is
demonstrated in `LeakPresenceDetectorTest.ResourceInStaticVariable1`.
This approach is pretty performant, but it requires changes in a few
places inside netty, and potentially in user code. Since this is only
necessary for the improved LeakPresenceDetector which is intended for
tests and opt-in, I think this is an acceptable compromise.

Alternative suggestions welcome.

Result:

Low-cost leak detection for tests. Bzip2IntegrationTest runs basically
just as fast as without leak presence detection.

Leaks can be unambiguously associated with a particular test, instead of
only being reported on GC.