Adaptive allocator tweaks for smaller heap sizes
Motivation:
With small heap sizes, we don't want to have as many chunks to reduce contention, and instead want higher utilization.

Modification:
Make the adaptive allocator detect when it's running in a "low memory" scenario, defined as having 512 MiB or less heap.
In that case, set the max stripe count to 1, disable the "large buffer" magazine group so "large" buffers are instead always unpooled, and disable the thread-local magazine groups.

Result:
We'll have fewer chunks with higher utilization when running in "low memory" scenarios.

In the allocation pattern simulator, we see some improvements.
Particularly when there's only a small number of live buffers:

| Live buffers | High mem | Low mem  |
|--------------|----------|----------|
| 128          | 5185536  | 4399104  |
| 512          | 9756672  | 8970240  |
| 1024         | 15671296 | 15165440 |