Make the ubuntu devcontainer use SDK MAN
Motivation:
This makes it easier to install arbitrary Java versions, as oppose to just the ones available through the apt package repository.

Modification:
Install SDK MAN into the ubuntu devcontainer and use it to install Java, instead of relying on apt for that.

Result:
We can now more easily switch to whichever Java version we want in the devcontainer, without being limited to what's available through apt.