Use exact byte length in `Unpooled#copiedBufferUtf8` (#15763)
Motivation:

The writeUtf8 method reserves the max number of bytes per char in the
underlying buffer. However we have the exact byte length of the utf char
sequence calculated already.

Modifications:

Use that exact byte length so the buffer does not need to grow.

Result:

The buffer will not need a grow on each operation.

Fixes #15751.

If there is no issue then describe the changes introduced by this PR.