Fix codec-native-quic Bundle-SymbolicNames
Motivation:

OSGi expects the combination of 'Bundle-SymblicName' and
'Bundle-Version' to be unique to a particular bundle, but we generate a
bundle for each jni.classifier.

Modification:

Run two generate-manifest executions, one for the main artifact and one
for native code. The second one overrides the usual Bundle-SymbolicName
by appending the jni.classifier (and includes the native-specific
headers).

Result:

Multiple CPU architectures can be supported by a single Apache Karaf
feature.

Signed-off-by: Robert Varga <robert.varga@pantheon.tech>