DnsNameResolver: Use inflight handling for all DnsQuestions (#15812)
Motivation:

At the moment we only did the inflight handling / merging for
resolve(hostname...) calls. We can do it for all DnsQuestion's and so
also use the optimization for others. This also fixes a bug where we did
not ignore the case of the hostname while doing so

Modifications:

Rewrite to make use of inflight handling / merging for all DnsQuestion's

Result:

More consistent handling