Add a filter when tasks are moved from scheduledTaskQueue to taskQueue. (#15826)
**Motivation:**
Currently, when fetching expired tasks from the delayed queue
(scheduledTaskQueue) and moving them to the task queue (taskQueue),
there’s no check to see if the tasks have been cancelled. This causes
cancelled tasks to be transferred to taskQueue, and the cancellation is
only detected when the tasks are actually executed.

**Modification:**
Add a check when fetching expired tasks from scheduledTaskQueue and
moving them to taskQueue.
Do not add tasks to taskQueue if they have been cancelled.

**Result:**
Reduced invalid operations such as the transfer of cancelled tasks
between queues and subsequent unnecessary processes.