Add close tracing to retain calls (#15832)
Motivation:
We attach close traces to release calls.
It is also possible that retain will be called on a released object, so
we should trace those as well.

Modification:
Attach close traces, if any, to any retaining methods that throw
IllegalReferenceCountException in SimpleLeakAwareByteBuf.

Result:
We now get close traces from various retain methods as well.