Guard against reentrancy in "normal" transports
Motivation:

It's possible that handlers / transports cause reentrancy in a ChannelHandler. This is problematic as it an result in various issues like data corruption etc. Due of this we should
try to guard against this as much as possible. While each handler could add its own guards it is kind of error-prone to do so.

Modification:

- Keep track of the current operation mode of the ChannelHandlerContext and so the ChannelHandler. If we are already executing an inbound operation and there is another one incoming just trampoline to the EventLoop to guard against reentrancy. The same is done for outbound operations.
- Enable the guard for every transport / channel except for EmbeddedChannel as trampolining does not work with this Channel implementation.

Result:

Guard against reentrancy issues for "real" transports