Enable X25519MLKEM768 by default for BoringSSL/AWS-LC (#15830)
Motivation:
The most urgent priority in post-quantum cryptography readiness, is
protecting the ephemeral session keys in TLS, because these are
vulnerable to harvest-now-decrypt-later attacks.
Thankfully, we can enable the hybrid-post-quantum X25519MLKEM768 key
exchange in a way that is backwards compatible and doesn't impact
connections to peers that don't support it.
We should do this by default, so we uplift the security of all our
downstream projects.
However, we can only know with certainty that this key exchange is
actually supported on BoringSSL and AWS-LC, so we only enable this by
default when our native TLS implementation is one of those.

Modification:
Add the X25519MLKEM768 group as first priority when using BoringSSL or
AWS-LC.
Add a test that proves that this is safe and interoperable to do.

Result:
Where supported by both peers, TLS connections will now use the
hybrid-post-quantum X25519MLKEM768 key exchange algorithm.