Remove ability to use different EventExecutorGroup for a ChannelHandler (#15889)
Motivation:

Remove deprecated API and feature of specify a different
EventExecutorGroup for a ChannelHandler. While this exists in earler
version of netty it complicated things a lot, introduced races and is
very error-prone. We decided it is much better to let the user offload
stuff by themselves.

Modifications:

Remove methods that take EventExecutorGroup in ChannelPipeline and
related code

Result:

Cleanup