Merge IoEventLoop into EventLoop and IoEventLoopGroup into EventLoopG… (#15905)
…roup

Motivation:

We introduced the IoEventLoop and IoEventLoopGroup interfaces into 4.2
to be able to rewrite how our EventLoops work in 4.2. This was just done
to not break the API, in 5.0 we should just merge both to simplify
stuff.

Modifications:

Merge EventLoop* interfaces

Result:

Cleanup and preparation for how Channels will be registered / assign to
an EventLoop