Tighten contract between Channel and EventLoop by require the EventLo… (#15910)
…op on Channel construction. 

Motivation:

At the moment it’s possible to have a Channel in Netty that is not
registered / assigned to an EventLoop until register(...) is called.
This is suboptimal as if the Channel is not registered it is also not
possible to do anything useful with a ChannelFuture that belongs to the
Channel. We should think about if we should have the EventLoop as a
constructor argument of a Channel and have the register / deregister
method only have the effect of add a Channel to KQueue/Epoll/... It is
also currently possible to deregister a Channel from one EventLoop and
register it with another EventLoop. This operation defeats the threading
model assumptions that are wide spread in Netty, and requires careful
user level coordination to pull off without any concurrency issues. It
is not a commonly used feature in practice, may be better handled by
other means (e.g. client side load balancing), and therefore we propose
removing this feature.

Modifications:

- Change all Channel implementations to require an EventLoop for
construction ( + an EventLoopGroup for all ServerChannel
implementations)
- Remove all register(...) methods from EventLoopGroup
- Change ChannelUnsafe.register(...) to not take an EventLoop as
parameter (as the EventLoop is supplied on custruction).
- Change ChannelFactory to take an EventLoop to create new Channels and
introduce ServerChannelFactory which takes an EventLoop and one
EventLoopGroup to create new ServerChannel instances.
- Add ServerChannel.childEventLoopGroup()
- Ensure all operations on the accepted Channel is done in the EventLoop
of the Channel in ServerBootstrap
- Change unit tests for new behaviour

Result:

A Channel always has an EventLoop assigned which will never change
during its life-time. This ensures we are always be able to call any
operation on the Channel once constructed (unit the EventLoop is
shutdown). This also simplifies the logic in DefaultChannelPipeline a
lot as we can always call handlerAdded / handlerRemoved directly without
the need to wait for register() to happen.

Also note that its still possible to deregister a Channel and register
it again. It's just not possible anymore to move from one EventLoop to
another (which was not really safe anyway).