Add ChannelOutboundInvoker.register(...) and ChannelOutboundHandler.r… (#15914)
…egister(...) methods

Motivation:

With the recent changes of taking the EventLoop as part of the
construction of Channel's we can let the whole register operation flow
through the pipeline as any other outbound operation. This will provide
more flexibility and is also more consistent with the rest.

Modifications:

- Add ChannelOutboundInvoker.register(...) and
ChannelOutboundHandler.register(...) methods
- Let the registration flow throught the ChannelPipeline

Result:

More consistent handling of outbound operations and more flexibility