Decouple AbstractChannel from DefaultChannelPipeline requirement (#15915)
Motivation:

AbstractChannel was tightly coupled to DefaultChannelPipeline before
which is really not necessary and also not a good design.
DefaultChannelPipeline is really just an implementation detail and
basically any ChannelPipeline should work.

Modification:

- Make use of the new ChannelOutboundHandler.register(...) method to do
the DefaultChannelPipeline specific code and so encapsulate it without
leaking anything into AbstractChannel
- Change return type of newChannelPipeline() method to ChannelPipeline

Result:

Less coupling and more flexibility