Remove method from Channel.Unsafe to retrieve ChannelOutboundBuffer (#15930)
Motivation:

ChannelOutboundBuffer is an implementation detail of AbstractChannel and
so should not leak into the public API. Let's remove it, this is a
followup of 8a5c3de402e3137005b8e5ce39164f303be91571.

Modifications:

- Remove method to retrieve ChannelOutboundBuffer from Unsafe.
- Add new Channel.hasPendingBytes() method and use it in SslHandler as
replacement of using ChannelOutboundBuffer
- Remove abilitiy to monitor output changes in IdleStateHandler as how
this was implemented was hacky and depended on ChannelOutboundBuffer.
- Disable traffic shapping handler for now as we need to find a better
way to support this.

Result:

Don't leak implementation details and also ensure people not mess with
internals