AbstractChannel: All outbound io operations must support asynchronous… (#15939)
… execution

Motivation:

AbstractChannel had various do* methods that sub-classes did override to
implement outbound io operations. As these methods did not take any
promise it was required for them to be non-async. This is quite limiting
and not really works for io_uring and so as a result we were not able to
replace all syscalls with io_uring ops. We should change this to make
our abstract base class more future proof and flexible.

Modifications:

- All do* methods for io outbound operations now take a promise that
needs to be notified once the operation is considered to be done.
- Adjust sub-classes for new method signature
- Remove hacks from io_uring impl as these are not needed anymore

Result:

Cleanup of API