Channel: Remove Channel.Unsafe access from public API and rename Channel.Unsafe to IoTransport
Motivation:

The whole Channel.Unsafe concept made it easy for people to mess up internals and so we shouldn't really allow them to access it. Because of this we should remove the access to it completely and just make it a concept for the DefaultChannelPipeline and AbstractChannel.

Modifications:

- Rename Channel.Unsafe and IoTransport and just have a final implementation of it in AbstractChannel
- Remove all the AbstractUnsafe extensions and just merge the methods into the Channel implementation itself.
- Rework IoHandle implementations to not extends AbstractUnsafe

Result:

Cleaner API and hiding internals from user