Merge *DomainSocket* implementations into other implementations (#15955)
Motiviation:

We should just merge our *DomainSocket* classes into the "normal"
Datagram* and *Socket* classes. By doing so we can eliminate a lot of
code and also reduce the class-hierarchy.

Modifications:

- Introduce SocketProtocolFamily and use it
- Merge Channel implementations

Result:

Code cleanup and easier to use.

This is basically a port of https://github.com/netty/netty/pull/12612