ChannelOutboundHandler: Remove throws Exception from method signature (#15974)
Motivation:

ChannelOutboundHandler methods take a ChannelPromise which should be
used to report errors and so it makes no sense to have throws Exception
on the method signature.

Modifications:

- Remove throws Exception from method signature
- Adjust code to correctly handle exceptions
- log warn message and close channel if an outbound operation throws an
exception but there is not way to report the error (flush and read).

Result:

API cleanup and make clear how errors should be reported back