ChannelHandler: Simplify ChannelHandler hierarchy
Motivation:

As we are now able to use default methods we can remove all the Channel*HandlerAdapter classes. This simplfies the hierachy a lot and reduced code that we need to maintain.
Beside this we can also get rid of the @Sharable annotation and just make isSharable() part of the ChannelHandler interface itself.

Modifications:

- Add default methods to ChannelHandler to remove ChannelHandlerAdapter
- Add default methods to ChannelInboundHandler to remove ChannelInboundHandlerAdapter
- Add default methods to ChannelOutboundHandler to remove ChannelOutboundHandlerAdapter
- Change ChannelDuplexHandler from a class to an interface
- Remove redunant code and just use the default methods where possible.
- Remove @Sharable and replace it with ChannelHandler.isSharable()
- Adjust code to use interfaces only and not the adapter classes.
- Remove all adapter classes.

Result:

Simplify ChannelHandler hierarchy and remove duplicated code