Use the XSI compliant strerror_r on Android when !_GNU_SOURCE
Motivation:

netty_unix_errors.c fails to build on Android platforms due to the
the current logic defaulting to the GNU strerror_r, which is not
exposed on Android when _GNU_SOURCE is not defined.

Android exposes the XSI strerror_r by default, so it can be used directly.
https://cs.android.com/android/platform/superproject/+/android16-release:bionic/libc/include/string.h;l=145?q=string.h

The GNU version is also available but only on API level 23+ (and
only when _GNU_SOURCE is defined), as explained at the link above.

Modifications:

- add __ANDROID__ to the platforms that use the XSI strerror_r when
  _GNU_SOURCE is not defined

Result:

netty_unix_errors.c compiles successfully on Android.
Tested on Android 16 emulator image, NDK 27.0.12077973