Make half-closure a core-concept of Channel (#15987)
Motivation:

At the moment our half-closure support is kind of weak and hacky. We use
various DuplexChannel specific methods which will not traverse the
pipeline at all and also use user events. This should be changed to be a
"core concept" and so have dedicated methods that can be used and also
traverse the pipeline.

What shutdown(...) actually means is a bit transport specific but in
general once someone shutdown the output it should not be possible to
write anymore. Also when the inbound is shutdown we should not propagate
inbound messages anymore. If the shutdown calls are propagated through
the network somehow is specific to the protocol itself and so might
differ.

Modifications:

- Add ChannelShutdownType and ChannelShutdownDirection
- Add ChannelOutboundInvoker.shutdown(ChannelShutdownType,
ChannelPromise) method
- Add ChannelInboundInvoker.fireChannelShutdown(ChannelShutdownType)
methods
- Add Channel.isShutdown(ChannelShutdownType) method
- Add ChannelHandler.shutdown(ChannelHandlerContext,
ChannelShutdownType) and
ChannelHandler.channelShutdown(ChannelHandlerContext,
ChannelShutdownType) methods
- Refactor code to use the new methods and also add implementations for
the different transports

Result:

Half-closure is now a core-concept of netty.